+++ src/lib/components/List/List.tsx
@@ -5,25 +5,26 @@ import styles from './List.module.scss';
 
 export interface IListProps {
 	children?: ReactElement<IListItemProps> | Array<ReactElement<IListItemProps>>;
+	className?: string;
 	isFlush?: boolean;
 	isHoverable?: boolean;
 }
 
 export const List = (props: IListProps) => {
 
-	const { children, isFlush = false, isHoverable = false } = props;
+	const { children, className = '', isFlush = false, isHoverable = false } = props;
 
-	const [listItems, setListItems] = useState<Array<ListItemModel>>([]);
+	// const [listItems, setListItems] = useState<Array<ListItemModel>>([]);
 
-	useEffect(() => {
-		if (children) {
-			if (Array.isArray(children)) {
-				setListItems(children.map(child => new ListItemModel(child)));
-			} else {
-				setListItems([new ListItemModel(children)]);
-			}
-		}
-	}, [children]);
+	// useEffect(() => {
+	// 	if (children) {
+	// 		if (Array.isArray(children)) {
+	// 			setListItems(children.map(child => new ListItemModel(child)));
+	// 		} else {
+	// 			setListItems([new ListItemModel(children)]);
+	// 		}
+	// 	}
+	// }, [children]);
 
 	const getCssClasses = () => {
 		const cssClasses: Array<string> = [];
@@ -31,17 +32,19 @@ export const List = (props: IListProps) => {
 			cssClasses.push(styles.flush);
 		}
 		cssClasses.push(styles.list);
+		cssClasses.push(className);
 		return cssClasses.filter(css => css).join(' ');
 	}
 
 	return (
 		<ul className={getCssClasses()}>
-			{listItems && listItems.map((listItem, index) => (
+			{children}
+			{/* {listItems && listItems.map((listItem, index) => (
 				cloneElement(listItem, {
 					isHoverable: isHoverable,
 					key: listItem.key ? listItem.key : index
 				})
-			))}
+			))} */}
 		</ul>
 	);
 }